requirements.txt
DateTime~=5.5
pyttsx3~=2.98
colorama~=0.4.6
pydub~=0.25.1
paho-mqtt~=2.1.0
paramiko~=3.5.0
mysql-connector-python~=9.1.0
aiohttp~=3.11.11
crypto~=1.4.1
pycryptodome~=3.21.0
requests~=2.32.3
pika~=1.3.2
PyAudio~=0.2.14
numpy~=2.2.1
future~=1.0.0
pystray~=0.19.5
pillow~=11.1.0
pygtail~=0.14.0
Flask~=3.1.0